Add Google Maps service for calculating distance Bill is now stored in the order and calculated in the create order endpoint

+++ src/Services/Orders/Orders.Application/Dtos/CostDetailDto.cs
@@ -5,6 +5,6 @@ public record CostDetailDto(
         Guid Id,
         Guid OrderId,
         string Description,
-        double Amount,
+        decimal Amount,
         string StatusC
     );
